+++ src/gcovr/data_model/coverage.py
@@ -135,11 +135,11 @@ class CoverageBase:
 
     __slots__ = ("data_sources",)
 
-    def __init__(self, data_source: Union[str, set[tuple[str, ...]]]) -> None:
-        if isinstance(data_source, str):
-            self.data_sources = set[tuple[str, ...]]([(data_source,)])
+    def __init__(self, data_sources: Union[str, set[tuple[str, ...]]]) -> None:
+        if isinstance(data_sources, str):
+            self.data_sources = set[tuple[str, ...]]([(data_sources,)])
         else:
-            self.data_sources = data_source
+            self.data_sources = data_sources
 
     def merge_base_data(
         self,
@@ -266,7 +266,7 @@ class BranchCoverage(CoverageBase):
     def __init__(
         self,
         parent: LineCoverage,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         branchno: Optional[int],
         count: int,
@@ -276,7 +276,7 @@ class BranchCoverage(CoverageBase):
         destination_block_id: Optional[int] = None,
         excluded: bool = False,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         self.branchno = branchno
         self.count = count
@@ -492,7 +492,7 @@ class ConditionCoverage(CoverageBase):
     def __init__(
         self,
         parent: LineCoverage,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         conditionno: int,
         count: int,
@@ -501,7 +501,7 @@ class ConditionCoverage(CoverageBase):
         not_covered_false: list[int],
         excluded: bool = False,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         self.conditionno = conditionno
         self.count = count
@@ -658,9 +658,9 @@ class DecisionCoverageUncheckable(CoverageBase):
     def __init__(
         self,
         parent: Optional[LineCoverage],
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
 
     def serialize(
@@ -726,12 +726,12 @@ class DecisionCoverageConditional(CoverageBase):
     def __init__(
         self,
         parent: Optional[LineCoverage],
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         count_true: int,
         count_false: int,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         if count_true < 0:
             self.raise_data_error("count_true must not be a negative value.")
@@ -817,11 +817,11 @@ class DecisionCoverageSwitch(CoverageBase):
     def __init__(
         self,
         parent: Optional[LineCoverage],
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         count: int,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         if count < 0:
             self.raise_data_error("count must not be a negative value.")
@@ -938,7 +938,7 @@ class CallCoverage(CoverageBase):
     def __init__(
         self,
         parent: LineCoverage,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         callno: Optional[int],
         source_block_id: int,
@@ -946,7 +946,7 @@ class CallCoverage(CoverageBase):
         returned: int,
         excluded: bool = False,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         self.callno = callno
         self.source_block_id = source_block_id
@@ -1113,8 +1113,6 @@ class LineCoverage(CoverageBase):
             List of block ids in this line
         excluded (bool, optional):
             Whether this line is excluded by a marker.
-        md5 (str, optional):
-            The md5 checksum of the source code line.
 
     >>> filecov = FileCoverage("file.gcov", filename="file.c")
     >>> linecov_list = LineCoverageCollection(filecov, "not_used.gcov", lineno=1)
@@ -1133,7 +1131,6 @@ class LineCoverage(CoverageBase):
         "demangled_function_name",
         "block_ids",
         "excluded",
-        "md5",
         "__branches",
         "__conditions",
         "decision",
@@ -1143,21 +1140,19 @@ class LineCoverage(CoverageBase):
     def __init__(
         self,
         parent: LineCoverageCollection,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         count: int,
         function_name: Optional[str],
         block_ids: Optional[list[int]] = None,
-        md5: Optional[str] = None,
         excluded: bool = False,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         self.count = count
         self.function_name = function_name
         self.demangled_function_name: Optional[str] = None
         self.block_ids = block_ids
-        self.md5 = md5
         self.excluded = excluded
         self.__branches = CoverageDict[BranchcovKeyType, BranchCoverage]()
         self.__conditions = CoverageDict[ConditioncovKeyType, ConditionCoverage]()
@@ -1273,8 +1268,6 @@ class LineCoverage(CoverageBase):
 
         Precondition: both objects must have same lineno.
         """
-        if self.lineno != other.lineno:
-            self.raise_merge_error("Line number must be equal.", other)
         if self.function_name != other.function_name:
             self.raise_merge_error("Function name must be equal.", other)
         # Merge the block_ids if present
@@ -1282,7 +1275,6 @@ class LineCoverage(CoverageBase):
             self.block_ids = other.block_ids
         elif other.block_ids is not None:
             self.block_ids = sorted(set(self.block_ids) | set(other.block_ids))
-        self.md5 = self._merge_property(other, "MD5 checksum", lambda x: x.md5)
 
         self.count += other.count
         self.excluded |= other.excluded
@@ -1324,7 +1316,7 @@ class LineCoverage(CoverageBase):
 
     def insert_branch_coverage(
         self,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         branchno: Optional[int],
         count: int,
@@ -1337,7 +1329,7 @@ class LineCoverage(CoverageBase):
         """Add a branch coverage item, merge if needed."""
         branchcov = BranchCoverage(
             self,
-            data_source,
+            data_sources,
             branchno=branchno,
             count=count,
             fallthrough=fallthrough,
@@ -1365,7 +1357,7 @@ class LineCoverage(CoverageBase):
 
     def insert_condition_coverage(
         self,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         conditionno: int,
         count: int,
@@ -1377,7 +1369,7 @@ class LineCoverage(CoverageBase):
         """Add a condition coverage item, merge if needed."""
         conditioncov = ConditionCoverage(
             self,
-            data_source=data_source,
+            data_sources=data_sources,
             conditionno=conditionno,
             count=count,
             covered=covered,
@@ -1405,7 +1397,7 @@ class LineCoverage(CoverageBase):
 
     def insert_call_coverage(
         self,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         callno: Optional[int],
         source_block_id: int,
@@ -1416,7 +1408,7 @@ class LineCoverage(CoverageBase):
         """Add a branch coverage item, merge if needed."""
         callcov = CallCoverage(
             self,
-            data_source=data_source,
+            data_sources=data_sources,
             callno=callno,
             source_block_id=source_block_id,
             destination_block_id=destination_block_id,
@@ -1446,6 +1438,11 @@ class LineCoverage(CoverageBase):
     def lineno(self) -> int:
         """Get the source location of the coverage data."""
         return self.parent.lineno
+
+    @property
+    def md5(self) -> Optional[str]:
+        """Get the md5of the source line."""
+        return self.parent.md5
 
     @property
     def is_excluded(self) -> bool:
@@ -1593,6 +1590,8 @@ class LineCoverageCollection(CoverageBase):
     Args:
         lineno (int):
             The line number.
+        md5 (str, optional):
+            The md5 checksum of the source code line.
 
     >>> filecov = FileCoverage("file.gcov", filename="file.c")
     >>> linecov_list = LineCoverageCollection(filecov, "line.gcov", lineno=0)
@@ -1603,39 +1602,128 @@ class LineCoverageCollection(CoverageBase):
        line.gcov
     """
 
-    __slots__ = ("parent", "lineno", "__linecovs")
+    __slots__ = ("parent", "lineno", "md5", "__linecov", "__raw_linecov")
 
     def __init__(
         self,
         parent: FileCoverage,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         lineno: int,
+        md5: Optional[str] = None,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         self.lineno = lineno
-        self.__linecovs = CoverageDict[LinecovKeyType, LineCoverage]()
+        self.md5 = md5
+        self.__linecov = CoverageDict[LinecovKeyType, LineCoverage]()
+        self.__raw_linecov = CoverageDict[LinecovKeyType, LineCoverage]()
 
         if lineno <= 0:
             self.raise_data_error("lineno must be a positive value.")
 
     def __setitem__(self, key: LinecovKeyType, item: LineCoverage) -> None:
-        self.__linecovs[key] = item
+        self.__linecov[key] = item
 
     def __getitem__(self, key: LinecovKeyType) -> LineCoverage:
-        return self.__linecovs[key]
+        return self.__linecov[key]
 
     def __len__(self) -> int:
-        return len(self.__linecovs)
+        return len(self.__linecov)
 
     def __iter__(self) -> Iterator[LineCoverage]:
-        return iter(self.__linecovs.values())
+        return iter(self.__linecov.values())
 
-    def linecovs(self) -> Iterable[LineCoverage]:
+    def linecov(self) -> Iterable[LineCoverage]:
         """Iterate over the lines."""
-        for _, linecov in sorted(self.__linecovs.items()):
+        for _, linecov in sorted(self.__linecov.items()):
             yield linecov
+
+    def raw_linecov(self) -> Iterable[LineCoverage]:
+        """Iterate over the lines."""
+        if self.__raw_linecov:
+            for _, linecov in sorted(self.__raw_linecov.items()):
+                yield linecov
+        else:
+            yield from self.linecov()
+
+    def merge_lines(self) -> None:
+        """Merge line coverage if there are several items for same line."""
+        if len(self) > 1:
+            # Remember the line coverage objects and clear the dictionary
+            separate_linecov = list(self)
+            self.__linecov.clear()
+            # Merge the information, needed for line coverage
+            data_sources = set[tuple[str, ...]]()
+            block_ids = set[int]()
+            for linecov in separate_linecov:
+                data_sources.update(linecov.data_sources)
+                if linecov.block_ids is not None:
+                    block_ids.update(linecov.block_ids)
+            new_linecov = self.insert_line_coverage(
+                data_sources,
+                count=sum(linecov.count for linecov in separate_linecov),
+                function_name=None,
+                block_ids=list(sorted(block_ids)) if block_ids else None,
+                excluded=all(linecov.is_excluded for linecov in separate_linecov),
+            )
+            # ...and add the child objects
+            for linecov in separate_linecov:
+                self.__raw_linecov[linecov.key] = linecov
+                for branchcov in linecov.branches():
+                    new_linecov.insert_branch_coverage(
+                        branchcov.data_sources,
+                        branchno=branchcov.branchno,
+                        count=branchcov.count,
+                        fallthrough=branchcov.fallthrough,
+                        throw=branchcov.throw,
+                        source_block_id=branchcov.source_block_id,
+                        destination_block_id=branchcov.destination_block_id,
+                        excluded=branchcov.excluded,
+                    )
+                for conditioncov in linecov.conditions():
+                    new_linecov.insert_condition_coverage(
+                        conditioncov.data_sources,
+                        conditionno=conditioncov.conditionno,
+                        count=conditioncov.count,
+                        covered=conditioncov.covered,
+                        not_covered_true=list(conditioncov.not_covered_true),
+                        not_covered_false=list(conditioncov.not_covered_false),
+                        excluded=conditioncov.excluded,
+                    )
+                if (decisioncov := linecov.decision) is not None:
+                    if isinstance(decisioncov, DecisionCoverageUncheckable):
+                        decisioncov = DecisionCoverageUncheckable(
+                            new_linecov,
+                            decisioncov.data_sources,
+                        )
+                    elif isinstance(decisioncov, DecisionCoverageConditional):
+                        decisioncov = DecisionCoverageConditional(
+                            new_linecov,
+                            decisioncov.data_sources,
+                            count_true=decisioncov.count_true,
+                            count_false=decisioncov.count_false,
+                        )
+                    elif isinstance(decisioncov, DecisionCoverageSwitch):
+                        decisioncov = DecisionCoverageSwitch(
+                            new_linecov,
+                            decisioncov.data_sources,
+                            count=decisioncov.count,
+                        )
+                    else:
+                        raise RuntimeError(
+                            "Sanity check failed, unknown decision type."
+                        )
+                    new_linecov.insert_decision_coverage(decisioncov)
+                for callcov in linecov.calls():
+                    new_linecov.insert_call_coverage(
+                        callcov.data_sources,
+                        callno=callcov.callno,
+                        source_block_id=callcov.source_block_id,
+                        destination_block_id=callcov.destination_block_id,
+                        returned=callcov.returned,
+                        excluded=callcov.excluded,
+                    )
 
     @property
     def key(self) -> LinecovCollectionKeyType:
@@ -1650,12 +1738,12 @@ class LineCoverageCollection(CoverageBase):
     @property
     def count(self) -> int:
         """Get the overall count of the line."""
-        return sum(linecov.count for linecov in self.linecovs())
+        return sum(linecov.count for linecov in self.linecov())
 
     @property
     def is_excluded(self) -> bool:
         """Return True if the line is excluded."""
-        return all(linecov.excluded for linecov in self.linecovs())
+        return all(linecov.excluded for linecov in self.linecov())
 
     @property
     def is_reportable(self) -> bool:
@@ -1674,51 +1762,49 @@ class LineCoverageCollection(CoverageBase):
 
     def exclude(self) -> None:
         """Exclude line from coverage statistic."""
-        for linecov in self.linecovs():
+        for linecov in self.linecov():
             linecov.excude()
 
     def decision_coverage(self) -> DecisionCoverageStat:
         """Return the decision coverage statistic of the line."""
         stat = DecisionCoverageStat.new_empty()
-        for linecov in self.linecovs():
+        for linecov in self.linecov():
             stat += linecov.decision_coverage()
         return stat
 
     def insert_line_coverage(
         self,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         options: MergeOptions = DEFAULT_MERGE_OPTIONS,
         *,
         count: int,
         function_name: Optional[str],
         block_ids: Optional[list[int]] = None,
-        md5: Optional[str] = None,
         excluded: bool = False,
     ) -> LineCoverage:
         """Add a line coverage item, merge if needed."""
         linecov = LineCoverage(
             self,
-            data_source,
+            data_sources,
             count=count,
             function_name=function_name,
             block_ids=block_ids,
-            md5=md5,
             excluded=excluded,
         )
         key = linecov.key
-        if key in self.__linecovs:
-            self.__linecovs[key].merge(linecov, options)
+        if key in self.__linecov:
+            self.__linecov[key].merge(linecov, options)
         else:
-            self.__linecovs[key] = linecov
-            self.__linecovs[key].parent = self
+            self.__linecov[key] = linecov
+            self.__linecov[key].parent = self
 
-        return self.__linecovs[key]
+        return self.__linecov[key]
 
     def remove_line_coverage(self, linecov: LineCoverage) -> None:
         """Remove line coverage object from line coverage collection."""
-        del self.__linecovs[linecov.key]
+        del self.__linecov[linecov.key]
         # Remove the line coverage collection if no data is available anymore for the line.
-        if not self.__linecovs:
+        if not self.__linecov:
             self.parent.remove_line(linecov.lineno)
 
     def merge(
@@ -1727,8 +1813,11 @@ class LineCoverageCollection(CoverageBase):
         options: MergeOptions,
     ) -> None:
         """Merge the data of the line coverage object."""
+        if self.lineno != other.lineno:
+            self.raise_merge_error("Line number must be equal.", other)
+        self.md5 = self._merge_property(other, "MD5 checksum", lambda x: x.md5)
         # pylint: disable=protected-access
-        self.__linecovs.merge(other.__linecovs, options)
+        self.__linecov.merge(other.__linecov, options)
         self.merge_base_data(other)
 
 
@@ -1795,7 +1884,7 @@ class FunctionCoverage(CoverageBase):
     def __init__(
         self,
         parent: FileCoverage,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         mangled_name: Optional[str],
         demangled_name: Optional[str],
@@ -1806,7 +1895,7 @@ class FunctionCoverage(CoverageBase):
         end: Optional[tuple[int, int]] = None,
         excluded: bool = False,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.parent = parent
         self.count = CoverageDict[int, int]({lineno: count})
         self.blocks = CoverageDict[int, float]({lineno: blocks})
@@ -2083,11 +2172,11 @@ class FileCoverage(CoverageBase):
 
     def __init__(
         self,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         *,
         filename: str,
     ) -> None:
-        super().__init__(data_source)
+        super().__init__(data_sources)
         self.filename: str = filename
         self.__functions = CoverageDict[FunctioncovKeyType, FunctionCoverage]()
         self.__lines = CoverageDict[LinecovCollectionKeyType, LineCoverageCollection]()
@@ -2103,9 +2192,9 @@ class FileCoverage(CoverageBase):
                     GCOVR_DATA_SOURCES: [
                         [
                             _presentable_filename(filename, options.root_filter)
-                            for filename in data_source
+                            for filename in data_sources
                         ]
-                        for data_source in sorted(cov.data_sources)
+                        for data_sources in sorted(cov.data_sources)
                     ]
                 }
         else:
@@ -2120,7 +2209,7 @@ class FileCoverage(CoverageBase):
         data_dict = {
             "file": filename,
             "lines": [
-                linecov.serialize(get_data_sources) for linecov in self.linecov()
+                linecov.serialize(get_data_sources) for linecov in self.raw_linecov()
             ],
             "functions": [
                 f
@@ -2135,7 +2224,7 @@ class FileCoverage(CoverageBase):
     @classmethod
     def deserialize(
         cls,
-        data_source: str,
+        data_sources: str,
         data_dict: dict[str, Any],
         merge_options: MergeOptions,
         options: Options,
@@ -2151,7 +2240,7 @@ class FileCoverage(CoverageBase):
         def get_data_sources(data_dict: dict[str, Any]) -> set[tuple[str, ...]]:
             """Return the set for data sources."""
             return set(
-                (*e,) for e in data_dict.get(GCOVR_DATA_SOURCES, [[data_source]])
+                (*e,) for e in data_dict.get(GCOVR_DATA_SOURCES, [[data_sources]])
             )
 
         filecov = FileCoverage(
@@ -2238,6 +2327,11 @@ class FileCoverage(CoverageBase):
             )
         del self.__lines[lineno]
 
+    def merge_lines(self) -> None:
+        """Merge line coverage if there are several items for same line."""
+        for linecov_collection in self.lines():
+            linecov_collection.merge_lines()
+
     def has_linecov(self) -> bool:
         """Test if there are line coverage objects available."""
         return any(linecov_collection for linecov_collection in self.lines())
@@ -2246,6 +2340,11 @@ class FileCoverage(CoverageBase):
         """Iterate over the line coverage objects."""
         for linecov_collection in self.lines():
             yield from linecov_collection
+
+    def raw_linecov(self) -> Iterable[LineCoverage]:
+        """Iterate over the line coverage objects."""
+        for linecov_collection in self.lines():
+            yield from linecov_collection.raw_linecov()
 
     @property
     def stats(self) -> SummarizedStats:
@@ -2261,7 +2360,7 @@ class FileCoverage(CoverageBase):
 
     def insert_line_coverage(
         self,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         options: MergeOptions = DEFAULT_MERGE_OPTIONS,
         *,
         lineno: int,
@@ -2272,7 +2371,9 @@ class FileCoverage(CoverageBase):
         excluded: bool = False,
     ) -> LineCoverage:
         """Add a line coverage item, merge if needed."""
-        linecov_collection = LineCoverageCollection(self, data_source, lineno=lineno)
+        linecov_collection = LineCoverageCollection(
+            self, data_sources, lineno=lineno, md5=md5
+        )
         key = linecov_collection.key
         if key in self.__lines:
             self.__lines[key].merge(linecov_collection, options)
@@ -2281,12 +2382,11 @@ class FileCoverage(CoverageBase):
             self.__lines[key].parent = self
 
         return self.__lines[key].insert_line_coverage(
-            data_source,
+            data_sources,
             options,
             count=count,
             function_name=function_name,
             block_ids=block_ids,
-            md5=md5,
             excluded=excluded,
         )
 
@@ -2296,7 +2396,7 @@ class FileCoverage(CoverageBase):
 
     def insert_function_coverage(
         self,
-        data_source: Union[str, set[tuple[str, ...]]],
+        data_sources: Union[str, set[tuple[str, ...]]],
         options: MergeOptions = DEFAULT_MERGE_OPTIONS,
         *,
         mangled_name: Optional[str],
@@ -2311,7 +2411,7 @@ class FileCoverage(CoverageBase):
         """Add a function coverage item, merge if needed."""
         functioncov = FunctionCoverage(
             self,
-            data_source,
+            data_sources,
             mangled_name=mangled_name,
             demangled_name=demangled_name,
             lineno=lineno,
@@ -2328,7 +2428,7 @@ class FileCoverage(CoverageBase):
             self.__functions[key] = functioncov
             self.__functions[key].parent = self
         for linecov_collection in self.lines():
-            for linecov in linecov_collection.linecovs():
+            for linecov in linecov_collection.linecov():
                 if functioncov.mangled_name is not None and (
                     functioncov.mangled_name == linecov.function_name
                 ):
@@ -2363,7 +2463,7 @@ class FileCoverage(CoverageBase):
             )
             filecov.__lines[lineno][linecov.key] = linecov
 
-        for linecov in self.linecov():
+        for linecov in self.raw_linecov():
             if functioncov.is_function(linecov.function_name):
                 add_linecov_to_collection(linecov)